+++ src/raw.rs
@@ -1,0 +1,1 @@
+pub struct Raw<T>(pub T);