+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## [Unreleased]
+- **Added**: Raw struct (to bypass escaping)
+- **Added**: DOCTYPE const (just a Raw for convenience)
 
 ## [0.2.0](https://codeberg.org/lsunsi/hiper/src/tag/0.2.0)
 - **Added**: 5-chars escaping to strings.